split `utils.test` into `random-/string-utils.test`
